tools/packages/build: If '${destdir}/share/autoconf/autoconf/trailer.m4' is present, export it as an environment variable.
When automake is being configured and build, it attempts to invoke 'autoconf'
to see if it works. In the context of this "sandboxed" autotools environment,
when invoked, 'autoconf' looks to the value of the environment variable
'trailer_m4', if set, and failing that '/share/autoconf/autoconf/trailer.m4'.
Since the latter is unlikely to ever exist (it would normally be
"/usr/share/autoconf/autoconf/trailer.m4"), by setting 'trailer_m4', we
ensure that this invocation test succeeds.